# contracts/openapi.yaml
openapi: "3.0.0"
info:
  version: "0.1.0"
  title: ContextPilot API
  description: |
    Meeting intelligence REST API.
    Phase 0 placeholder: metadata and common schemas only.
    Endpoints added in BRD-02 and later domain specs.
servers:
  - url: /api/v1
    description: Base API path
...
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
  /meetings:
    get:
      operationId: listMeetings
      summary: List all meetings
      description: Returns an array of Meeting objects with nested participants.
      tags:
        - meetings
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        '200':
          description: List of meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Meeting'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      operationId: createMeeting
      summary: Create a completed meeting record
      description: Creates a new completed meeting with structured participants. Gated by FF_ENABLE_MANUAL_MEETING_IMPORT.
      tags:
        - meetings
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeetingCreate'
      responses:
        '201':
          description: Meeting created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  redirect:
                    type: string
                    format: uri
                    example: /meetings/550e8400-e29b-41d4-a716-446655440000
        '400':
          description: Validation failure
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                          example: title
                        message:
                          type: string
                          example: Title is required and must be at most 500 characters.
                  values:
                    type: object
                    description: Submitted values for form re-rendering
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Feature flag disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Idempotency token already consumed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum: [duplicate]
                  meetingId:
                    type: string
                    format: uuid
                  originalOutcome:
                    type: string
                    example: completed
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  correlationId:
                    type: string
                    format: uuid
...
  /meetings/{id}:
    get:
      operationId: getMeeting
      summary: Get a meeting by ID
      description: Returns a single Meeting object with nested participants.
      tags:
        - meetings
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Meeting details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meeting'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Meeting not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    patch:
      operationId: updateMeeting
      summary: Update a meeting
      description: Updates a completed meeting record. Triggers stale memory detection for BRD-03. Gated by FF_ENABLE_MANUAL_MEETING_IMPORT.
      tags:
        - meetings
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeetingInput'
      responses:
        '200':
          description: Meeting updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meeting'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Meeting not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
...
  schemas:
    Problem:
      description: RFC 7807 error response
      type: object
      properties:
        type:
          type: string
          format: uri
          example: https://contextpilot.app/errors/problem
        title:
          type: string
          example: Resource Not Found
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: Meeting with ID 12345 does not exist or you do not have access.
        instance:
          type: string
          format: uri
          example: /api/v1/meetings/12345
        request_id:
          type: string
...
    Participant:
      type: object
      description: A structured meeting participant attached to a meeting record.
      required:
        - id
        - displayName
        - displayOrder
      properties:
        id:
          type: string
          format: uuid
        meetingId:
          type: string
          format: uuid
          nullable: true
          description: ID of the parent meeting. Present when listed under meeting_participants.
        displayName:
          type: string
          maxLength: 500
          description: Display name as entered; leading/trailing whitespace trimmed.
        email:
          type: string
          format: email
          nullable: true
        organization:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
        displayOrder:
          type: integer
          description: Sort order within the meeting participant list.
    Meeting:
      type: object
      description: A completed meeting record with structured participants.
      required:
        - id
        - title
        - completedAt
        - meeting_participants
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 500
        completedAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the meeting was completed.
        transcript:
          type: string
          nullable: true
          description: Full transcript text, null when notes-only.
        notes:
          type: string
          nullable: true
          description: Meeting notes, null when transcript-only.
        meeting_participants:
          type: array
          items:
            $ref: "#/components/schemas/Participant"
        contentSource:
          type: string
          enum: [transcript, notes, both]
          description: Derived at save time based on which content fields are populated.
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
          description: User ID of the meeting creator.
        displayOrder:
          type: integer
          description: Sort order for display purposes.
    MeetingCreate:
      type: object
      description: Request body for creating a new completed meeting record.
      required:
        - title
        - completedAt
        - participants
        - idempotencyToken
      properties:
        title:
          type: string
          maxLength: 500
        completedAt:
          type: string
          format: date-time
        participants:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - displayName
            properties:
              displayName:
                type: string
                maxLength: 500
              email:
                type: string
                format: email
              organization:
                type: string
              role:
                type: string
        transcript:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        idempotencyToken:
          type: string
          format: uuid
          description: UUID generated on form load; dedupes repeated saves within 24h.
    UpdateMeetingInput:
      type: object
      description: Request body for updating a completed meeting record. All fields optional — only provided fields are updated.
      properties:
        title:
          type: string
          maxLength: 500
        completedAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the meeting was completed.
        transcript:
          type: string
          nullable: true
          description: Full transcript text.
        notes:
          type: string
          nullable: true
          description: Meeting notes.
        participants:
          type: array
          description: Full replacement participant list. Omit to keep existing participants.
          items:
            type: object
            required:
              - displayName
            properties:
              displayName:
                type: string
                maxLength: 500
              email:
                type: string
                format: email
                nullable: true
              organization:
                type: string
                nullable: true
              role:
                type: string
                nullable: true
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: session

  # ─── Memory Processing Schemas ───────────────────────────────────────────

    MemoryProcessingState:
      type: object
      description: Current processing lifecycle state for a meeting's memory.
      properties:
        meeting_id:
          type: string
          format: uuid
        processing_state:
          type: string
          enum: [not_processed, queued, processing, completed, completed_with_insufficient_evidence, failed, stale, reprocessing, retrying, retry_exhausted]
        active_version_number:
          type: integer
          nullable: true
          description: Present when a memory version exists.
        is_stale:
          type: boolean
          description: True when source meeting has been edited after the active version was created.
        latest_job:
          type: object
          nullable: true
          properties:
            job_id:
              type: string
              format: uuid
            status:
              type: string
              enum: [queued, processing, completed, completed_with_insufficient_evidence, failed, retrying, retry_exhausted]
            retry_count:
              type: integer
...
  securitySchemes: